requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14

# Vector database
chromadb==0.4.22

# PDF processing
pdfplumber==0.10.3

# Embeddings and ML
sentence-transformers>=3.0.0
torch>=2.0.0
numpy<2.0.0

# Utilities
pydantic>=2.9.0
python-dotenv==1.0.0

# LLM Integration
openai>=1.0.0
